Create blaseburnout
I made this for determining the amount of burnout steps, it works pretty well. Basically it compares last value of the magnetism history along with the middle value(50 percent of array) against the amount of allowed error we set. Then it adds or subtracts powers of 1/2 to the 50 percent and compares the value at that new location with the last value of magnetism history again, and repeats this process. This doesn't work for extremely noisy magnetism history, in which case the burnout steps doesn't matter anyway.